requirements.txt
torch
torchvision
numpy
scikit-learn
opencv-python
pybloom-live
datasets
Pillow
